# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Siroko Senior Challenge API
  version: 1.0.0
  description: |
    API REST para el Siroko Senior PHP Challenge.

    Integra dos bounded contexts:
    - **Catalog + Cart + Order** (Opción A): gestión de productos, carrito y checkout.
    - **Semantic Search** (Opción B): búsqueda por intención con embeddings y fallback a LLM.

    ### Autenticación
    Los endpoints de escritura del catálogo requieren autenticación mediante API Key:
    ```
    Authorization: Bearer <ADMIN_API_KEY>
    ```
    El resto de la API es pública.
servers:
  - url: http://localhost:8080
    description: Docker local
...
security: []  # por defecto sin auth; los endpoints admin lo declaran explícitamente
components:
  securitySchemes:
    AdminApiKey:
      type: http
      scheme: bearer
      description: API Key estática configurada en ADMIN_API_KEY

  # -------------------------------------------------------------------------
  # Schemas reutilizables
  # -------------------------------------------------------------------------
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: integer
          description: Importe en céntimos (nunca float)
          example: 9999
        currency:
          type: string
          enum: [EUR, USD, GBP]
          example: EUR
      example:
        amount: 9999
        currency: EUR
    Category:
      type: string
      enum: [CYCLING, FITNESS, APPAREL, ACCESSORIES]
      example: CYCLING
...
    ProductAttributes:
      type: object
      description: Atributos extensibles del producto (colores, tallas, materiales…)
      properties:
        colors:
          type: array
          items:
            type: string
          example: ["negro", "azul marino"]
        sizes:
          type: array
          items:
            type: string
          example: ["S", "M", "L", "XL"]
        materials:
          type: array
          items:
            type: string
          example: ["lycra", "poliéster"]
      additionalProperties:
        type: array
        items:
          type: string

    # --- Product -------------------------------------------------------------
    ProductDTO:
      type: object
      required: [id, name, description, price, category, brand, stock, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "Mallot Siroko K38 Carbon"
        description:
          type: string
          example: "Mallot de ciclismo de alto rendimiento con tejido termorregulador."
        price:
          $ref: '#/components/schemas/Money'
        category:
          $ref: '#/components/schemas/Category'
        brand:
          type: string
          example: "Siroko"
        attributes:
          $ref: '#/components/schemas/ProductAttributes'
        stock:
          type: integer
          minimum: 0
          example: 42
        imageUrl:
          type: string
          format: uri
          nullable: true
          example: "https://cdn.siroko.com/products/k38-carbon.jpg"
        createdAt:
          type: string
          format: date-time
          example: "2026-05-28T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-28T10:00:00Z"
    CreateProductRequest:
      type: object
      required: [name, price, category, brand, stock]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "Mallot Siroko K38 Carbon"
        description:
          type: string
          maxLength: 2000
          default: ""
          example: "Mallot de ciclismo de alto rendimiento."
        price:
          $ref: '#/components/schemas/Money'
        category:
          $ref: '#/components/schemas/Category'
        brand:
          type: string
          minLength: 1
          maxLength: 100
          example: "Siroko"
        attributes:
          $ref: '#/components/schemas/ProductAttributes'
        stock:
          type: integer
          minimum: 0
          example: 42
        imageUrl:
          type: string
          format: uri
          nullable: true
          example: "https://cdn.siroko.com/products/k38-carbon.jpg"
...
    ProductListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductDTO'
        pagination:
          type: object
          required: [page, pageSize, total, totalPages]
          properties:
            page:
              type: integer
              example: 1
            pageSize:
              type: integer
              example: 20
            total:
              type: integer
              example: 143
            totalPages:
              type: integer
              example: 8

    # --- Cart ----------------------------------------------------------------
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: product_not_found
            message:
              type: string
              example: "Product with id '123e4567-...' was not found."
            context:
              type: object
              additionalProperties: true
              example:
                productId: "123e4567-e89b-12d3-a456-426614174000"

  # -------------------------------------------------------------------------
  # Reusable responses
  # -------------------------------------------------------------------------
  responses:
    Unauthorized:
      description: Token ausente o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthorized
              message: "Invalid API key."
...
    UnprocessableEntity:
      description: Error de validación o regla de negocio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/products:
    post:
      summary: Crear producto
      description: |
        Crea un nuevo producto en el catálogo.
        La indexación semántica en Qdrant se realiza automáticamente en background
        mediante el evento `ProductCreated`.
      tags: [Catalog (admin)]
      security:
        - AdminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
            example:
              name: "Mallot Siroko K38 Carbon"
              description: "Mallot de ciclismo de alto rendimiento con tejido termorregulador."
              price:
                amount: 9999
                currency: EUR
              category: CYCLING
              brand: "Siroko"
              attributes:
                colors: ["negro", "azul marino"]
                sizes: ["S", "M", "L", "XL"]
              stock: 42
              imageUrl: "https://cdn.siroko.com/products/k38-carbon.jpg"
      responses:
        '201':
          description: Producto creado
          content:
            application/json:
              schema:
                type: object
                required: [productId]
                properties:
                  productId:
                    $ref: '#/components/schemas/UUID'
              example:
                productId: "123e4567-e89b-12d3-a456-426614174000"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    get:
      summary: Listar productos
      tags: [Catalog (public)]
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/Category'
        - name: brand
          in: query
          schema:
            type: string
          example: Siroko
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
              example:
                data:
                  - id: "123e4567-e89b-12d3-a456-426614174000"
                    name: "Mallot Siroko K38 Carbon"
                    description: "Mallot de ciclismo de alto rendimiento."
                    price:
                      amount: 9999
                      currency: EUR
                    category: CYCLING
                    brand: Siroko
                    attributes:
                      colors: ["negro", "azul marino"]
                      sizes: ["S", "M", "L", "XL"]
                    stock: 42
                    imageUrl: "https://cdn.siroko.com/products/k38-carbon.jpg"
                    createdAt: "2026-05-28T10:00:00Z"
                    updatedAt: "2026-05-28T10:00:00Z"
                pagination:
                  page: 1
                  pageSize: 20
                  total: 143
                  totalPages: 8
...